requirements.txt
Flask==2.3.3
openai==1.35.0
python-dotenv==1.0.0
WTForms==3.0.1
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
httpx==0.24.1